add de grabado de recicladorUrbano

+++ src/main/java/org/circle8/controller/request/user/UserRequest.java
@@ -18,8 +18,8 @@ public class UserRequest implements IRequest {
 
 	/* Opcionales */
 	public String razonSocial;
-	public Integer zonaId;
-	public Integer organizacionId;
+	public Long zonaId;
+	public Long organizacionId;
 
 	@Override
 	public Validation valid() {
